edit and update department

+++ app/Http/Controllers/DepartmentController.php
@@ -65,7 +65,12 @@ class DepartmentController extends Controller
      */
     public function edit($id)
     {
-        //
+        $department = Department::find($id);
+
+        if (isset($department))
+            return view('editdepartment')->with('department', $department);
+
+        return redirect('/departments');
     }
 
     /**
@@ -77,7 +82,15 @@ class DepartmentController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        $department = Department::find($id);
+
+        if (isset($department)) {
+            $department->name = $request->input('departmentName');
+            $department->save();
+        }
+            
+
+        return redirect('/departments');
     }
 
     /**
@@ -90,8 +103,8 @@ class DepartmentController extends Controller
     {
         $department = Department::find($id);
 
-        
-        $department->delete();
+        if (isset($department))
+            $department->delete();
 
         return redirect('/departments');
     }
